+++ e2e/discoverSheetFlow.spec.js
@@ -154,5 +154,6 @@ describe('Discover Sheet Flow', () => {
   afterAll(async () => {
     // Reset the app state
     await device.clearKeychain();
+    await Helpers.delay(2000);
   });
 });