+++ e2e/ganacheTransactionFlow.spec.js
@@ -10,20 +10,6 @@ let uri = null;
 let account = null;
 
 beforeAll(async () => {
-  // Create WC client
-  // Create a connector
-  connector = new WalletConnect({
-    bridge: 'https://zora.bridge.walletconnect.org',
-    clientMeta: {
-      description: 'Connect with WalletConnect',
-      icons: ['https://walletconnect.org/walletconnect-logo.png'],
-      name: 'WalletConnect',
-      url: 'https://walletconnect.org',
-    },
-  });
-
-  await connector.createSession();
-  uri = connector.uri;
   // Connect to ganache
   await exec('yarn ganache');
 });
@@ -102,40 +88,39 @@ describe('Ganache Transaction Flow', () => {
     await Helpers.swipe('profile-screen', 'left', 'slow');
   });
 
-  /*
-  it('Should swap ETH -> ERC20 (DAI)', async () => {
-    await Helpers.tap('exchange-fab');
-    await Helpers.typeText('exchange-modal-input', '0.01', true);
-    await Helpers.tap('exchange-modal-output-selection-button');
-    await Helpers.typeText('currency-select-search-input', 'DAI', true);
-    await Helpers.tap('exchange-coin-row-DAI');
-    await Helpers.tapAndLongPress('exchange-modal-confirm');
-    await Helpers.swipe('profile-screen', 'left', 'slow');
-  });
+  // it('Should swap ETH -> ERC20 (DAI)', async () => {
+  //   await Helpers.tap('exchange-fab');
+  //   await Helpers.typeText('exchange-modal-input', '0.01', true);
+  //   await Helpers.tap('exchange-modal-output-selection-button');
+  //   await Helpers.typeText('currency-select-search-input', 'DAI', true);
+  //   await Helpers.tap('exchange-coin-row-DAI');
+  //   await Helpers.tapAndLongPress('exchange-modal-confirm');
+  //   await Helpers.swipe('profile-screen', 'left', 'slow');
+  // });
 
-  it('Should swap ERC20 (BAT) -> ERC20 (ZRX)', async () => {
-    await Helpers.tap('exchange-fab');
-    await Helpers.tap('exchange-modal-input-selection-button');
-    await Helpers.tap('exchange-coin-row-BAT');
-    await Helpers.typeText('exchange-modal-input', '5', true);
-    await Helpers.tap('exchange-modal-output-selection-button');
-    await Helpers.tap('exchange-coin-row-ZRX');
-    await Helpers.tapAndLongPress('exchange-modal-confirm');
-    await Helpers.swipe('profile-screen', 'left', 'slow');
-  });
+  // it('Should swap ERC20 (BAT) -> ERC20 (ZRX)', async () => {
+  //   await Helpers.tap('exchange-fab');
+  //   await Helpers.tap('exchange-modal-input-selection-button');
+  //   await Helpers.tap('exchange-coin-row-BAT');
+  //   await Helpers.typeText('exchange-modal-input', '5', true);
+  //   await Helpers.tap('exchange-modal-output-selection-button');
+  //   await Helpers.tap('exchange-coin-row-ZRX');
+  //   await Helpers.tapAndLongPress('exchange-modal-confirm');
+  //   await Helpers.swipe('profile-screen', 'left', 'slow');
+  // });
 
-  it('Should swap ERC20 (USDC)-> ETH', async () => {
-    await Helpers.tap('exchange-fab');
-    await Helpers.tap('exchange-modal-input-selection-button');
-    await Helpers.tap('exchange-coin-row-USDC');
-    await Helpers.typeText('exchange-modal-input', '2', true);
-    await Helpers.tap('exchange-modal-output-selection-button');
-    await Helpers.typeText('currency-select-search-input', 'ETH', true);
-    await Helpers.tap('exchange-coin-row-ETH');
-    await Helpers.tapAndLongPress('exchange-modal-confirm');
-    await Helpers.swipe('profile-screen', 'left', 'slow');
-  });
-*/
+  // it('Should swap ERC20 (USDC)-> ETH', async () => {
+  //   await Helpers.tap('exchange-fab');
+  //   await Helpers.tap('exchange-modal-input-selection-button');
+  //   await Helpers.tap('exchange-coin-row-USDC');
+  //   await Helpers.typeText('exchange-modal-input', '2', true);
+  //   await Helpers.tap('exchange-modal-output-selection-button');
+  //   await Helpers.typeText('currency-select-search-input', 'ETH', true);
+  //   await Helpers.tap('exchange-coin-row-ETH');
+  //   await Helpers.tapAndLongPress('exchange-modal-confirm');
+  //   await Helpers.swipe('profile-screen', 'left', 'slow');
+  // });
+
   it('Should send ERC20 (cSAI)', async () => {
     await Helpers.tap('send-fab');
     await Helpers.typeText('send-asset-form-field', 'poopcoin.eth', false);
@@ -188,10 +173,22 @@ describe('Ganache Transaction Flow', () => {
     await Helpers.tap('send-sheet-confirm-action-button');
     await Helpers.tapAndLongPress('send-confirmation-button');
     await Helpers.checkIfVisible('profile-screen');
-    await Helpers.swipe('profile-screen', 'left', 'slow');
   });
 
   it('Should receive the WC connect request and approve it', async () => {
+    connector = new WalletConnect({
+      bridge: 'https://bridge.walletconnect.org',
+      clientMeta: {
+        description: 'Connect with WalletConnect',
+        icons: ['https://walletconnect.org/walletconnect-logo.png'],
+        name: 'WalletConnect',
+        url: 'https://walletconnect.org',
+      },
+    });
+    await Helpers.delay(3000);
+
+    await connector.createSession();
+    uri = connector.uri;
     const connected = new Promise(async (resolve, reject) => {
       connector.on('connect', (error, payload) => {
         if (error) {
@@ -210,11 +207,15 @@ describe('Ganache Transaction Flow', () => {
     const baseUrl = 'https://rnbwapp.com';
     const encodedUri = encodeURIComponent(uri);
     const fullUrl = `${baseUrl}/wc?uri=${encodedUri}`;
+
+    await Helpers.disableSynchronization();
     await device.sendToHome();
+    await Helpers.enableSynchronization();
+
+    await Helpers.delay(2000);
 
     await device.launchApp({
       newInstance: false,
-      sourceApp: 'com.apple.mobilesafari',
       url: fullUrl,
     });
 
@@ -323,7 +324,6 @@ describe('Ganache Transaction Flow', () => {
       throw new Error('WC approving tx failed');
     }
     await Helpers.delay(3000);
-    await Helpers.swipe('wallet-screen', 'right', 'slow');
   });
 
   /*
@@ -347,6 +347,7 @@ describe('Ganache Transaction Flow', () => {
       await Helpers.checkIfVisible('Sending-Compound Sai');
     }
   });*/
+
   it('Should show completed send ERC20 (cSAI)', async () => {
     try {
       await Helpers.checkIfVisible('Sent-Compound SAI-1.69 cSAI');
@@ -395,5 +396,6 @@ describe('Ganache Transaction Flow', () => {
     connector = null;
     await device.clearKeychain();
     await exec('kill $(lsof -t -i:7545)');
+    await Helpers.delay(2000);
   });
 });