+++ src/components/expanded-state/profile/ProfileAvatarButton.js
@@ -1,32 +1,21 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { ButtonPressAnimation } from '../../animations';
 import { ContactAvatar } from '../../contacts';
 
 const ProfileAvatarButton = ({
+  changeAvatar,
   color,
   marginBottom = 15,
-  setColor,
   testID,
   value,
   radiusAndroid,
 }) => {
-  const { colors } = useTheme();
-  const handleChangeColor = useCallback(() => {
-    setColor?.(prevColor => {
-      const idx =
-        typeof prevColor === 'string'
-          ? colors.avatarColor.indexOf(prevColor)
-          : prevColor;
-      return (idx + 1) % colors.avatarColor.length;
-    });
-  }, [setColor, colors]);
-
   return (
     <ButtonPressAnimation
-      onPress={handleChangeColor}
+      onPress={changeAvatar}
       overflowMargin={15}
       radiusAndroid={radiusAndroid}
-      scaleTo={0.96}
+      scaleTo={0.9}
       testID={testID}
     >
       <ContactAvatar