+++ src/components/transaction-list/TransactionList.js
@@ -1,6 +1,6 @@
 import Clipboard from '@react-native-community/clipboard';
 import analytics from '@segment/analytics-react-native';
-import { startCase, toLower } from 'lodash';
+import { pick, startCase, toLower } from 'lodash';
 import React, { useCallback, useMemo, useRef, useState } from 'react';
 import { requireNativeComponent } from 'react-native';
 import { useDispatch } from 'react-redux';
@@ -95,7 +95,14 @@ export default function TransactionList({
     });
   }, [navigate, isDamaged]);
 
-  const onAvatarPress = useOnAvatarPress();
+  const {
+    avatarOptions,
+    onAvatarChooseImage,
+    onAvatarRemovePhoto,
+    onAvatarPickEmoji,
+    onAvatarPress,
+    onAvatarWebProfile,
+  } = useOnAvatarPress();
 
   const onReceivePress = useCallback(() => {
     if (isDamaged) {
@@ -256,14 +263,6 @@ export default function TransactionList({
     navigate(Routes.CHANGE_WALLET_SHEET);
   }, [navigate]);
 
-  const data = useMemo(
-    () => ({
-      requests,
-      transactions,
-    }),
-    [requests, transactions]
-  );
-
   const loading = useMemo(() => (!initialized && !isFocused()) || isLoading, [
     initialized,
     isLoading,
@@ -275,6 +274,50 @@ export default function TransactionList({
   const safeAccountImage = useSafeImageUri(accountImage);
   const { isDarkMode, colors } = useTheme();
 
+  const onNativeAvatarMenuSelect = useCallback(
+    e => {
+      const { selection } = e.nativeEvent;
+      switch (selection) {
+        case 'newimage':
+          onAvatarChooseImage();
+          break;
+        case 'newemoji':
+          onAvatarPickEmoji();
+          break;
+        case 'removeimage':
+          onAvatarRemovePhoto();
+          break;
+        case 'webprofile':
+          onAvatarWebProfile();
+          break;
+        default:
+          break;
+      }
+    },
+    [
+      onAvatarChooseImage,
+      onAvatarPickEmoji,
+      onAvatarRemovePhoto,
+      onAvatarWebProfile,
+    ]
+  );
+
+  const data = useMemo(() => {
+    const requestsNative = requests.map(request =>
+      pick(request, [
+        'clientId',
+        'dappName',
+        'imageUrl',
+        'payloadId',
+        'displayDetails.timestampInMs',
+      ])
+    );
+    return {
+      requests: requestsNative,
+      transactions,
+    };
+  }, [requests, transactions]);
+
   return (
     <Container>
       <Container
@@ -284,6 +327,7 @@ export default function TransactionList({
         accountName={accountSymbol}
         addCashAvailable={addCashAvailable}
         as={NativeTransactionListView}
+        avatarOptions={avatarOptions}
         darkMode={isDarkMode}
         data={data}
         isAvatarPickerAvailable={isAvatarPickerAvailable}
@@ -292,6 +336,7 @@ export default function TransactionList({
         onAddCashPress={onAddCashPress}
         onAvatarPress={onAvatarPress}
         onCopyAddressPress={onCopyAddressPress}
+        onNativeAvatarMenuSelect={onNativeAvatarMenuSelect}
         onReceivePress={onReceivePress}
         onRequestExpire={onRequestExpire}
         onRequestPress={onRequestPress}