+++ src/hooks/useWalletConnectConnections.js
@@ -4,9 +4,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import { createSelector } from 'reselect';
 import { sortList } from '../helpers/sortList';
 import {
-  walletConnectDisconnectAllByPeerId as rawWalletConnectDisconnectAllByPeerId,
+  walletConnectDisconnectAllByDappUrl as rawWalletConnectDisconnectAllByDappUrl,
   walletConnectOnSessionRequest as rawWalletConnectOnSessionRequest,
-  walletConnectUpdateSessionConnectorByPeerId as rawWalletConnectUpdateSessionConnectorByPeerId,
+  walletConnectUpdateSessionConnectorByDappUrl as rawWalletConnectUpdateSessionConnectorByDappUrl,
 } from '../redux/walletconnect';
 
 const formatDappData = connections =>
@@ -42,8 +42,8 @@ export default function useWalletConnectConnections() {
     walletConnectorsCount,
   } = useSelector(walletConnectSelector);
 
-  const walletConnectDisconnectAllByPeerId = useCallback(
-    peerId => dispatch(rawWalletConnectDisconnectAllByPeerId(peerId)),
+  const walletConnectDisconnectAllByDappUrl = useCallback(
+    dappUrl => dispatch(rawWalletConnectDisconnectAllByDappUrl(dappUrl)),
     [dispatch]
   );
 
@@ -53,11 +53,11 @@ export default function useWalletConnectConnections() {
     [dispatch]
   );
 
-  const walletConnectUpdateSessionConnectorByPeerId = useCallback(
-    (peerId, accountAddress, chainId) =>
+  const walletConnectUpdateSessionConnectorByDappUrl = useCallback(
+    (dappUrl, accountAddress, chainId) =>
       dispatch(
-        rawWalletConnectUpdateSessionConnectorByPeerId(
-          peerId,
+        rawWalletConnectUpdateSessionConnectorByDappUrl(
+          dappUrl,
           accountAddress,
           chainId
         )
@@ -67,10 +67,10 @@ export default function useWalletConnectConnections() {
 
   return {
     sortedWalletConnectors,
-    walletConnectDisconnectAllByPeerId,
+    walletConnectDisconnectAllByDappUrl,
     walletConnectOnSessionRequest,
     walletConnectorsByDappName,
     walletConnectorsCount,
-    walletConnectUpdateSessionConnectorByPeerId,
+    walletConnectUpdateSessionConnectorByDappUrl,
   };
 }