+++ src/redux/usersPositions.ts
@@ -1,3 +1,4 @@
+import { isEmpty } from 'lodash';
 import { AnyAction } from 'redux';
 import { uniswapClient } from '../apollo/client';
 import {
@@ -5,6 +6,10 @@ import {
   USER_MINTS_BURNS_PER_PAIR,
   USER_POSITIONS,
 } from '../apollo/queries';
+import {
+  getUniswapPositions,
+  saveUniswapPositions,
+} from '@rainbow-me/handlers/localstorage/uniswap';
 import { AppDispatch, AppGetState } from '@rainbow-me/redux/store';
 import {
   BUSD_ADDRESS,
@@ -34,12 +39,12 @@ export interface TypeSpecificParameters {
   supplyBalanceUnderlying: string;
 }
 
-interface PositionsState {
-  [key: string]: any;
+export interface PositionsState {
+  [key: string]: UniswapPosition[];
 }
 
 interface ReturnMetrics {
-  hodleReturn: number; // difference in asset values t0 -> t1 with t0 deposit amounts
+  hodlReturn: number; // difference in asset values t0 -> t1 with t0 deposit amounts
   netReturn: number; // net return from t0 -> t1
   uniswapReturn: number; // netReturn - hodlReturn
   impLoss: number;
@@ -58,13 +63,28 @@ interface Position {
   token1PriceUSD: number;
 }
 
-export type StoredPositions = Position &
+export type UniswapPosition = Position &
   ReturnMetrics & { fees: { sum: number } };
 
 // --- fetching ----------------//
 
 // -- Constants --------------------------------------- //
 const UPDATE_POSITIONS = 'positions/UPDATE_POSITIONS';
+
+export const uniswapPositionsLoadState = () => async (
+  dispatch: AppDispatch,
+  getState: AppGetState
+) => {
+  const { accountAddress, network } = getState().settings;
+  try {
+    const positions = await getUniswapPositions(accountAddress, network);
+    dispatch({
+      payload: positions,
+      type: UPDATE_POSITIONS,
+    });
+    // eslint-disable-next-line no-empty
+  } catch (error) {}
+};
 
 function formatPricesForEarlyTimestamps(position: any): Position {
   if (position.timestamp < PRICE_DISCOVERY_START_TIMESTAMP) {
@@ -149,7 +169,7 @@ function getMetricsForPositionWindow(
 
   return {
     fees: difference_fees_usd,
-    hodleReturn: assetValueT1 - assetValueT0,
+    hodlReturn: assetValueT1 - assetValueT0,
     impLoss: imp_loss_usd,
     netReturn: netValueT1 - netValueT0,
     uniswapReturn: uniswap_return,
@@ -257,7 +277,7 @@ async function getLPReturnsOnPair(
         : snapshots[parseInt(index) + 1];
 
     const results = getMetricsForPositionWindow(positionT0, positionT1);
-    hodlReturn = hodlReturn + results.hodleReturn;
+    hodlReturn = hodlReturn + results.hodlReturn;
     netReturn = netReturn + results.netReturn;
     uniswapReturn = uniswapReturn + results.uniswapReturn;
     fees = fees + results.fees;
@@ -305,7 +325,7 @@ async function fetchSnapshots(account: string): Promise<Position[]> {
   return [];
 }
 
-async function fetchData(account: string): Promise<StoredPositions[]> {
+async function fetchData(account: string): Promise<UniswapPosition[]> {
   const priceOfEther = ethereumUtils.getEthPriceUnit();
 
   try {
@@ -342,17 +362,25 @@ async function fetchData(account: string): Promise<StoredPositions[]> {
 
 // -- Actions ---------------------------------------- //
 
-export const updatePositions = async (
+export const updatePositions = () => async (
   dispatch: AppDispatch,
   getState: AppGetState
 ) => {
-  const { accountAddress } = getState().settings;
+  const { accountAddress, network } = getState().settings;
+  const existingPositions = getState().usersPositions;
   const data = await fetchData(accountAddress);
+  if (!isEmpty(data)) {
+    const payload = {
+      ...existingPositions,
+      [accountAddress]: data,
+    };
 
-  dispatch({
-    payload: { [accountAddress]: data },
-    type: UPDATE_POSITIONS,
-  });
+    dispatch({
+      payload,
+      type: UPDATE_POSITIONS,
+    });
+    saveUniswapPositions(payload, accountAddress, network);
+  }
 };
 
 const INITIAL_STATE: PositionsState = {};